# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is docs of Celestia indexer API.",
        "title": "Swagger Celestia Indexer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1",
    "basePath": "/v1",
    "paths": {
        "/v1/address": {
            "get": {
                "description": "List address info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "List address info",
                "operationId": "list-address",
                "parameters": [
                    {
                        "maximum": 100,
                        "type": "integer",
                        "description": "Count of requested entities",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Address"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/address/{hash}": {
            "get": {
                "description": "Get address info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "address"
                ],
                "summary": "Get address info",
                "operationId": "get-address",
                "parameters": [
                    {
                        "maxLength": 48,
                        "minLength": 48,
                        "type": "string",
                        "description": "Hash",
                        "name": "hash",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Address"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/block": {
            "get": {
                "description": "List blocks info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "List blocks info",
                "operationId": "list-block",
                "parameters": [
                    {
                        "maximum": 100,
                        "type": "integer",
                        "description": "Count of requested entities",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "Sort order",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Block"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/block/{height}": {
            "get": {
                "description": "Get block info",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Get block info",
                "operationId": "get-block",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Block height",
                        "name": "height",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.Block"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/v1/block/{height}/events": {
            "get": {
                "description": "Get events from begin and end of block",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "block"
                ],
                "summary": "Get events from begin and end of block",
                "operationId": "get-block-events",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Block height",
                        "name": "height",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.Event"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "responses.Address": {
            "description": "Celestia address information",
            "type": "object",
            "properties": {
                "balance": {
                    "type": "string",
                    "example": "10000000000"
                },
                "first_height": {
                    "type": "integer",
                    "example": 100
                },
                "hash": {
                    "type": "string",
                    "example": "celestia1jc92qdnty48pafummfr8ava2tjtuhfdw774w60"
                },
                "id": {
                    "type": "integer",
                    "example": 321
                }
            }
        },
        "responses.Block": {
            "type": "object",
            "properties": {
                "app_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "consensus_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "data_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "events_count": {
                    "type": "integer",
                    "example": 18
                },
                "evidence_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "height": {
                    "type": "integer",
                    "example": 100
                },
                "id": {
                    "type": "integer",
                    "example": 321
                },
                "last_commit_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "last_results_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "message_types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "MsgSend",
                        "MsgUnjail"
                    ]
                },
                "next_validators_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "parent_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "proposer_address": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "time": {
                    "type": "string",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "tx_count": {
                    "type": "integer",
                    "example": 12
                },
                "validators_hash": {
                    "type": "string",
                    "example": "652452A670018D629CC116E510BA88C1CABE061336661B1F3D206D248BD558AF"
                },
                "version_app": {
                    "type": "string",
                    "example": "1"
                },
                "version_block": {
                    "type": "string",
                    "example": "11"
                }
            }
        },
        "responses.Event": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "height": {
                    "type": "integer",
                    "format": "int64",
                    "example": 100
                },
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 321
                },
                "position": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "time": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2023-07-04T03:10:57+00:00"
                },
                "tx_id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 11
                },
                "type": {
                    "type": "string",
                    "format": "string",
                    "enum": [
                        "coin_received",
                        "coinbase",
                        "coin_spent",
                        "burn",
                        "mint",
                        "message",
                        "proposer_reward",
                        "rewards",
                        "commission",
                        "liveness",
                        "attestation_request",
                        "transfer",
                        "pay_for_blobs",
                        "redelegate",
                        "withdraw_rewards",
                        "withdraw_commission",
                        "create_validator",
                        "delegate",
                        "edit_validator",
                        "unbond",
                        "tx",
                        "unknown"
                    ],
                    "example": "commission"
                }
            }
        },
...
    }
}